pyproject.toml
[tool.poetry]
name = "asteroidHunter"
version = "0.1.0"
description = ""
authors = ["Chandu <crotta@corp.lawyer.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
